Puts Record in alignment.rs
