+++ src/alignment.rs
@@ -769,6 +769,58 @@ impl BaseAlignment {
     }
 }
 
+#[pymethods]
+impl Record {
+    #[new]
+    /// Creates a new Record object from sequence_id, sequence_description
+    /// and sequence_str.
+    fn __new__(obj: &PyRawObject, id: &str, description: &str, sequence_str: &str) -> PyResult<()> {
+        let sequence_str = String::from(sequence_str);
+        obj.init(|_| {
+            Record {
+                id: id.to_string(),
+                description: description.to_string(),
+                sequence: sequence_str.to_string(),
+            }
+        })
+    }
+    fn sequence_to_uint32(&self) -> Vec<i32> {
+        let uints: Vec<i32> = self.sequence.chars().map(|x| {x as usize} as i32 ).collect();
+        uints
+    }
+}
+
+// Customizes __repr__ and __str__ of PyObjectProtocol trait
+#[pyproto]
+impl PyObjectProtocol for Record {
+    fn __repr__(&self) -> PyResult<String> {
+        let desc:String = match self.description.chars().count() {
+            x if x > 20 => {
+                let mut desc:String = self.description.char_indices()
+                                        .filter(|(i, _)| *i < 17)
+                                        .map(|(_, c)| c)
+                                        .collect();
+                desc.push_str("...");
+                desc
+            },
+            _ => self.description.clone()
+        };
+        Ok(format!("Record(id=\"{id}\", len={seq_len}, description=\"{desc}\")", id=self.id, seq_len=self.sequence.len(), desc=desc))
+    }
+
+    fn __str__(&self) -> PyResult<String> {
+        if self.description.len() > 0 {
+            return Ok(format!(">{id} {desc}\n{seq_len}",
+                id=self.id,
+                desc=self.description,
+                seq_len=self.sequence.len()))
+        }
+        return Ok(format!(">{id}\n{seq_len}",
+                id=self.id,
+                seq_len=self.sequence.len()))        
+    }
+}
+
 #[pyfunction]
 /// Concatenates a list of alignments over the site (column) axis.
 fn concat_basealignments(aln_list: Vec<&BaseAlignment>) -> PyResult<BaseAlignment> {
@@ -796,8 +848,8 @@ fn concat_basealignments(aln_list: Vec<&BaseAlignment>) -> PyResult<BaseAlignmen
 fn alignment(_py: Python, m: &PyModule) -> PyResult<()> {
     m.add_function(wrap_function!(concat_basealignments))?;
 
-    // Add Block class
     m.add_class::<BaseAlignment>()?;
+    m.add_class::<Record>()?;
 
     Ok(())
 }
